added modules to organize animals
