+++ Slithering/__init__.py
@@ -1,0 +1,5 @@
+from .copperheadSlithering import CopperHead
+from .gardensnakeSlithering import GardenSnake
+from .pythonSlithering import Python
+from .ratsnakeSlithering import RatSnake
+from .viperSlithering import Viper
