+++ Slithering/slithering.py
@@ -1,0 +1,57 @@
+from datetime import date
+
+class CopperHead:
+
+    def __init__(self, name , species):
+        # Establish the properties of each animal
+        # with a default value
+        self.name = name
+        self.species = species
+        self.date_added = date.today()
+        self.area = "Glass Tank"
+        self.slithering = True
+
+class Python:
+
+    def __init__(self, name , species):
+        # Establish the properties of each animal
+        # with a default value
+        self.name = name
+        self.species = species
+        self.date_added = date.today()
+        self.area = "Glass Tank"
+        self.slithering = True
+
+class RatSnake:
+
+    def __init__(self, name , species):
+        # Establish the properties of each animal
+        # with a default value
+        self.name = name
+        self.species = species
+        self.date_added = date.today()
+        self.area = "Glass Tank"
+        self.slithering = True
+
+
+class GardenSnake:
+
+    def __init__(self, name , species):
+        # Establish the properties of each animal
+        # with a default value
+        self.name = name
+        self.species = species
+        self.date_added = date.today()
+        self.area = "Glass Tank"
+        self.slithering = True
+        
+class Viper:
+
+    def __init__(self, name , species):
+        # Establish the properties of each animal
+        # with a default value
+        self.name = name
+        self.species = species
+        self.date_added = date.today()
+        self.area = "Glass Tank"
+        self.slithering = True
