+++ Slithering/viperSlithering.py
@@ -1,0 +1,12 @@
+from datetime import date
+
+class Viper:
+
+    def __init__(self, name , species):
+        # Establish the properties of each animal
+        # with a default value
+        self.name = name
+        self.species = species
+        self.date_added = date.today()
+        self.area = "Glass Tank"
+        self.slithering = True
