+++ Swimming/__init__.py
@@ -1,0 +1,5 @@
+from .clownfishSwimming import ClownFish
+from .goldfishSwimming import GoldFish
+from .koifishSwimming import KoiFish
+from .malladsSwimming import Mallards
+from .troutSwimming import Trout
