+++ Swimming/clownfishSwimming.py
@@ -1,0 +1,12 @@
+from datetime import date
+
+class ClownFish:
+
+    def __init__(self, name , species):
+        # Establish the properties of each animal
+        # with a default value
+        self.name = name
+        self.species = species
+        self.date_added = date.today()
+        self.area = "Pond"
+        self.swimming = True
