+++ Walking/__init__.py
@@ -1,0 +1,5 @@
+from .cowWalking import Cow
+from .donkeyWalking import Donkey
+from .goatWalking import Goat
+from .llamaWalking import Llama
+from .rabbitWalking import Rabbit
