--- animals.py
@@ -1,203 +1,0 @@
-# import the python datetime module to help us create a timestamp
-from datetime import date
-
-class Llama:
-
-    def __init__(self):
-        # Establish the properties of each animal
-        # with a default value
-        self.name = ""
-        self.species = ""
-        self.date_added = date.today()
-        self.area = "Petting Area"
-        self.walking = True
-        
-class Donkey:
-
-    def __init__(self):
-        # Establish the properties of each animal
-        # with a default value
-        self.name = "mr. donkey"
-        self.species = ""
-        self.date_added = date.today()
-        self.area = "Petting Area"
-        self.walking = True
-        
-class Goat:
-
-    def __init__(self):
-        # Establish the properties of each animal
-        # with a default value
-        self.name = ""
-        self.species = ""
-        self.date_added = date.today()
-        self.area = "Petting Area"
-        self.walking = True
-        
-class Cow:
-
-    def __init__(self):
-        # Establish the properties of each animal
-        # with a default value
-        self.name = ""
-        self.species = ""
-        self.date_added = date.today()
-        self.area = "Petting Area"
-        self.walking = True
-      
-class Rabbit:
-
-    def __init__(self):
-        # Establish the properties of each animal
-        # with a default value
-        self.name = ""
-        self.species = ""
-        self.date_added = date.today()
-        self.area = "Petting Area"
-        self.walking = True
-
-class CopperHead:
-
-    def __init__(self):
-        # Establish the properties of each animal
-        # with a default value
-        self.name = ""
-        self.species = ""
-        self.date_added = date.today()
-        self.area = "Glass Tank"
-        self.slithering = True
-
-class Python:
-
-    def __init__(self):
-        # Establish the properties of each animal
-        # with a default value
-        self.name = ""
-        self.species = ""
-        self.date_added = date.today()
-        self.area = "Glass Tank"
-        self.slithering = True
-
-class RatSnake:
-
-    def __init__(self):
-        # Establish the properties of each animal
-        # with a default value
-        self.name = ""
-        self.species = ""
-        self.date_added = date.today()
-        self.area = "Glass Tank"
-        self.slithering = True
-
-
-class GardenSnake:
-
-    def __init__(self):
-        # Establish the properties of each animal
-        # with a default value
-        self.name = ""
-        self.species = ""
-        self.date_added = date.today()
-        self.area = "Glass Tank"
-        self.slithering = True
-        
-class Viper:
-
-    def __init__(self):
-        # Establish the properties of each animal
-        # with a default value
-        self.name = ""
-        self.species = ""
-        self.date_added = date.today()
-        self.area = "Glass Tank"
-        self.slithering = True
-
-class GoldFish:
-
-    def __init__(self):
-        # Establish the properties of each animal
-        # with a default value
-        self.name = ""
-        self.species = ""
-        self.date_added = date.today()
-        self.area = "Pond"
-        self.swimming = True
-        
-class Mallards:
-
-    def __init__(self):
-        # Establish the properties of each animal
-        # with a default value
-        self.name = ""
-        self.species = ""
-        self.date_added = date.today()
-        self.area = "Pond"
-        self.swimming = True
-        
-class KoiFish:
-
-    def __init__(self):
-        # Establish the properties of each animal
-        # with a default value
-        self.name = ""
-        self.species = ""
-        self.date_added = date.today()
-        self.area = "Pond"
-        self.swimming = True
-        
-class ClownFish:
-
-    def __init__(self):
-        # Establish the properties of each animal
-        # with a default value
-        self.name = ""
-        self.species = ""
-        self.date_added = date.today()
-        self.area = "Pond"
-        self.swimming = True
-    
-class Trout:
-
-    def __init__(self):
-        # Establish the properties of each animal
-        # with a default value
-        self.name = ""
-        self.species = ""
-        self.date_added = date.today()
-        self.area = "Pond"
-        self.swimming = True
-
-
-mr_donkey = Donkey()
-print(mr_donkey.name)
-
-cowcow = Cow()
-
-ribbonrabbit = Rabbit()
-
-miss_fuzz = Llama()
-
-supergoat = Goat()
-
-copper = CopperHead()
-
-ratman = RatSnake()
-
-mrcoder = Python()
-
-mrsviper = Viper()
-
-gardenman = GardenSnake()
-
-goldswimmer = GoldFish()
-
-mallman = Mallards()
-
-coins = KoiFish()
-
-orangefish = ClownFish()
-
-John = Trout()
-
-miss_fuzz.name = "Miss Fuzz"
-miss_fuzz.species = "domestic llama"
