+++ index.py
@@ -1,0 +1,47 @@
+from Slithering import CopperHead, GardenSnake, Python, RatSnake, Viper
+from Swimming import ClownFish, GoldFish, KoiFish, Mallards, Trout
+from Walking import Cow, Donkey, Goat, Llama, Rabbit
+
+
+
+
+
+
+
+
+
+
+# import the python datetime module to help us create a timestamp
+# from datetime import date
+
+# mr_donkey = Donkey()
+# print(mr_donkey.name)
+
+# cowcow = Cow()
+
+# ribbonrabbit = Rabbit()
+
+# miss_fuzz = Llama()
+
+# supergoat = Goat()
+
+# copper = CopperHead()
+
+# ratman = RatSnake()
+
+# mrcoder = Python()
+
+# mrsviper = Viper()
+
+# gardenman = GardenSnake()
+
+# goldswimmer = GoldFish()
+
+# mallman = Mallards()
+
+# coins = KoiFish()
+
+# orangefish = ClownFish()
+
+# John = Trout()
+
